Coverage uses native HTML generation when coverage 3 is installed (Issue 264)
Thanks to Timothee Peignier for the patch. This also removes the verbose
doctest that was testing HTML output because the new HTML format is a lot more
complicated.  Ideally we should add tests that parse the generated HTML to
spot check it. Instead there are just some basic sanity checks.